# docs/openapi.yaml
openapi: 3.0.3

info:
  title: EventHub — Event Management System API
  version: "1.0.0"
  description: |
    REST API for the Event Management System built with Node.js + Express + MongoDB.
    
    **Authentication:** Bearer JWT token. Obtain via `/api/auth/login` then include as:
    ```
    Authorization: Bearer <token>
    ```
    
    **Roles:**
    - `admin` — Full access
    - `organizer` — Create/manage own events
    - `attendee` — Browse events and register
  contact:
    name: IILM University Project Team
    email: team@eventhub.com

servers:
  - url: http://localhost:5000/api
    description: Local development
  - url: https://eventhub-api.onrender.com/api
    description: Production (Render)

tags:
  - name: Auth
    description: Authentication and user profile
  - name: Events
    description: Event CRUD and management
  - name: Registrations
    description: Registration and ticketing
  - name: Admin
    description: Admin-only dashboard and user management

# ─────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────
paths:

  # ── AUTH ──────────────────────────────────────────────────
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:     { type: string, example: Sadique Khan }
                email:    { type: string, format: email, example: sadique@iilm.ac.in }
                password: { type: string, minLength: 6, example: secret123 }
                role:     { type: string, enum: [attendee, organizer], default: attendee }
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "400":
          $ref: '#/components/responses/BadRequest'

  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, example: sadique@iilm.ac.in }
                password: { type: string, example: secret123 }
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'

  /auth/me:
    get:
      tags: [Auth]
      summary: Get logged-in user profile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      tags: [Auth]
      summary: Update own profile
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:   { type: string }
                phone:  { type: string }
                avatar: { type: string }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ── EVENTS ───────────────────────────────────────────────
  /events:
    get:
      tags: [Events]
      summary: List all published events (public)
      parameters:
        - in: query
          name: category
          schema: { type: string, enum: [conference, workshop, seminar, festival] }
        - in: query
          name: search
          schema: { type: string }
          description: Search by title (regex)
        - in: query
          name: status
          schema: { type: string, enum: [published, ongoing] }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Paginated list of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      events: { type: array, items: { $ref: '#/components/schemas/Event' } }
                      total:  { type: integer }
                      page:   { type: integer }
                      pages:  { type: integer }

    post:
      tags: [Events]
      summary: Create a new event
      security: [{ bearerAuth: [] }]
      description: Accessible by organizers and admins only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        "403":
          $ref: '#/components/responses/Forbidden'

  /events/{id}:
    get:
      tags: [Events]
      summary: Get a single event by ID (public)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        "404":
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Events]
      summary: Update an event
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'

    delete:
      tags: [Events]
      summary: Delete an event
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Deleted successfully
        "403":
          $ref: '#/components/responses/Forbidden'

  /events/{id}/publish:
    patch:
      tags: [Events]
      summary: Publish a draft event
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Event published

  /events/{id}/registrations:
    get:
      tags: [Events]
      summary: Get all registrations for an event
      security: [{ bearerAuth: [] }]
      description: Accessible by the event's organizer or admin only.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Registration'

  # ── REGISTRATIONS ─────────────────────────────────────────
  /registrations:
    post:
      tags: [Registrations]
      summary: Register for an event
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId]
              properties:
                eventId:    { type: string, example: 65a1b2c3d4e5f6789abc0010 }
                ticketType: { type: string, enum: [free, paid, vip], default: free }
      responses:
        "201":
          description: Registration confirmed with QR code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        "400":
          description: Already registered / Event full

  /registrations/my:
    get:
      tags: [Registrations]
      summary: Get logged-in user's own registrations
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of user's registrations

  /registrations/{id}:
    get:
      tags: [Registrations]
      summary: Get a specific registration
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Registration details

  /registrations/{id}/cancel:
    patch:
      tags: [Registrations]
      summary: Cancel a registration
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Registration cancelled

  /registrations/{id}/checkin:
    patch:
      tags: [Registrations]
      summary: Check in an attendee at the event
      security: [{ bearerAuth: [] }]
      description: Only organizers and admins can check in attendees.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Attendee checked in

  # ── ADMIN ─────────────────────────────────────────────────
  /admin/stats:
    get:
      tags: [Admin]
      summary: Get system-wide dashboard statistics
      security: [{ bearerAuth: [] }]
      description: Admin only.
      responses:
        "200":
          description: Aggregated stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalEvents:         { type: integer }
                  totalUsers:          { type: integer }
                  totalRegistrations:  { type: integer }
                  recentRegistrations: { type: array, items: { $ref: '#/components/schemas/Registration' } }

  /admin/users:
    get:
      tags: [Admin]
      summary: List all users
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: All users

  /admin/users/{id}/role:
    patch:
      tags: [Admin]
      summary: Change a user's role
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [admin, organizer, attendee] }
      responses:
        "200":
          description: Role updated

  /admin/users/{id}:
    delete:
      tags: [Admin]
      summary: Delete a user
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: User deleted

# ─────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────
components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: MongoDB ObjectId

  schemas:

    User:
      type: object
      properties:
        _id:       { type: string }
        name:      { type: string }
        email:     { type: string }
        role:      { type: string, enum: [admin, organizer, attendee] }
        phone:     { type: string }
        avatar:    { type: string }
        isActive:  { type: boolean }
        createdAt: { type: string, format: date-time }

    AuthResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            token: { type: string }
            user:  { $ref: '#/components/schemas/User' }

    Session:
      type: object
      properties:
        _id:         { type: string }
        title:       { type: string }
        speaker:     { type: string }
        startTime:   { type: string, format: date-time }
        endTime:     { type: string, format: date-time }
        location:    { type: string }
        description: { type: string }

    Event:
      type: object
      properties:
        _id:         { type: string }
        title:       { type: string }
        description: { type: string }
        date:        { type: string, format: date-time }
        endDate:     { type: string, format: date-time }
        venue:       { type: string }
        capacity:    { type: integer }
        category:    { type: string, enum: [conference, workshop, seminar, festival] }
        status:      { type: string, enum: [draft, published, ongoing, completed, cancelled] }
        organizer:   { $ref: '#/components/schemas/User' }
        sessions:    { type: array, items: { $ref: '#/components/schemas/Session' } }
        tags:        { type: array, items: { type: string } }
        createdAt:   { type: string, format: date-time }

    EventInput:
      type: object
      required: [title, date, venue, capacity, category]
      properties:
        title:       { type: string, example: TechConf 2026 }
        description: { type: string }
        date:        { type: string, format: date-time }
        endDate:     { type: string, format: date-time }
        venue:       { type: string, example: Auditorium A }
        capacity:    { type: integer, minimum: 1, example: 300 }
        category:    { type: string, enum: [conference, workshop, seminar, festival] }
        sessions:    { type: array, items: { $ref: '#/components/schemas/Session' } }
        tags:        { type: array, items: { type: string } }

    Registration:
      type: object
      properties:
        _id:         { type: string }
        user:        { $ref: '#/components/schemas/User' }
        event:       { $ref: '#/components/schemas/Event' }
        ticketType:  { type: string, enum: [free, paid, vip] }
        qrCode:      { type: string, description: Base64 data URI of QR ticket }
        status:      { type: string, enum: [pending, confirmed, cancelled] }
        checkInTime: { type: string, format: date-time, nullable: true }
        createdAt:   { type: string, format: date-time }

  responses:
    BadRequest:
      description: Bad request — validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              message: { type: string }
    Unauthorized:
      description: Unauthorized — invalid or missing token
    Forbidden:
      description: Forbidden — insufficient role permissions
    NotFound:
      description: Resource not found